Move breakpoints into grid component
