--- client-react/src/design/breakpoint-columns.js
@@ -1,14 +1,0 @@
-export default {
-  default: 12,
-  4000: 11,
-  3600: 10,
-  3200: 9,
-  2800: 8,
-  2400: 7,
-  2000: 6,
-  1700: 5,
-  1400: 4,
-  1100: 3,
-  800: 2,
-  500: 1,
-};
