+++ client-react/src/pages/Register/index.js
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import TileAvatar from '../../components/TileAvatar';
 import TileFormRegister from '../../components/TileFormRegister';
 import Grid from '../../components/Grid';
-import breakpointColumns from '../../design/breakpoint-columns';
 import avatars from '../../shared/avatars';
 import chat from '../../shared/chat';
 import Logo from '../../assets/images/mediaan-logo.png';
@@ -19,7 +18,7 @@ const Register = () => {
   };
 
   return (
-    <Grid breakpointCols={breakpointColumns}>
+    <Grid>
       <TileImage image={Logo} alt="Mediaan" />
       <Tile appearance={chocolate}>
         <StyledTitle>Charlie Chatlin</StyledTitle>
