+++ client-react/src/service-worker.js
